Refactor download button and export utils
#440
