+++ src/lineup/internal/ExportUtils.ts
@@ -1,39 +1,7 @@
-import {IDataRow, Column, isNumberColumn, LocalDataProvider, isSupportType, isDateColumn, UIntTypedArray} from 'lineupjs';
-import {BaseUtils, I18nextManager} from 'phovea_core';
+import {IDataRow, Column, isNumberColumn, isDateColumn} from 'lineupjs';
 import {XlsxUtils} from '../../utils/XlsxUtils';
 
 export declare type ExportFormat = 'json' | 'csv' | 'tsv' | 'ssv' | 'xlsx';
-export declare type ExportRows = 'all' | 'filtered' | 'selected';
-
-/**
- * Store the ordered row indices
- */
-export interface IOrderedRowIndices {
-  /**
-   * All row indices from the data provider.
-   * Indices are not sorting (= sorting of input data)!
-   */
-  all: number[];
-
-  /**
-   * Indices of the selected rows.
-   * Indices are sorted by the *first* ranking.
-   */
-  selected: number[];
-
-  /**
-   * Indices of the filtered rows.
-   * Indices are sorted and filtered by the *first* ranking.
-   */
-  filtered: number[];
-}
-
-interface IExportData {
-  type: ExportFormat;
-  columns: Column[];
-  order: number[];
-  name: string;
-}
 
 export class ExportUtils {
 
@@ -76,7 +44,7 @@ export class ExportUtils {
     return JSON.stringify(converted, null, 2);
   }
 
-  static exportxlsx(columns: Column[], rows: IDataRow[]) {
+  static exportXLSX(columns: Column[], rows: IDataRow[]) {
     const converted = rows.map((row) => {
       const r: any = {};
       for (const col of columns) {
@@ -90,143 +58,6 @@ export class ExportUtils {
         columns: columns.map((d) => ({name: ExportUtils.getColumnName(d), type: <'float' | 'string' | 'date'>(isNumberColumn(d) ? 'float' : isDateColumn(d) ? 'date' : 'string')})),
         rows: converted
       }]
-    });
-  }
-
-  static exportLogic(format: 'custom' | ExportFormat, rows: 'custom' | ExportRows, orderedRowIndices: IOrderedRowIndices, provider: LocalDataProvider) {
-    if (format === 'custom') {
-      return ExportUtils.customizeDialog(orderedRowIndices, provider).then((r) => ExportUtils.convertRanking(provider, r.order, r.columns, r.type, r.name));
-
-    } else {
-      const ranking = provider.getFirstRanking();
-      let order: number[];
-
-      switch(rows) {
-        case 'selected':
-          order = orderedRowIndices.selected;
-          break;
-        case 'filtered':
-          order = orderedRowIndices.filtered;
-          break;
-        default:
-          order = orderedRowIndices.all;
-      }
-
-      const columns = ranking.flatColumns.filter((c) => !isSupportType(c));
-      return Promise.resolve(ExportUtils.convertRanking(provider, order, columns, format, ranking.getLabel()));
-    }
-  }
-
-  private static toBlob(content: string, mimeType: string) {
-    return new Blob([content], {type: mimeType});
-  }
-
-  private static convertRanking(provider: LocalDataProvider, order: number[], columns: Column[], type: ExportFormat, name: string) {
-    const rows = provider.viewRawRows(order);
-    const separators = {csv: ',', tsv: '\t', ssv: ';'};
-    let content: Promise<Blob> | Blob;
-    const mimeTypes = {csv: 'text/csv', tsv: 'text/tab-separated-values', ssv: 'text/csv', json: 'application/json', xlsx: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'};
-    const mimeType = mimeTypes[type];
-    if (type in separators) {
-      content = ExportUtils.toBlob(ExportUtils.exportRanking(columns, rows, separators[type]), mimeType);
-    } else if (type === 'xlsx') {
-      content = ExportUtils.exportxlsx(columns, rows);
-    } else { // json
-      content = ExportUtils.toBlob(ExportUtils.exportJSON(columns, rows), mimeType);
-    }
-    return Promise.resolve(content).then((c) => ({
-      content: c,
-      mimeType: mimeTypes[type],
-      name: `${name}.${type === 'ssv' ? 'csv' : type}`
-    }));
-  }
-
-  private static customizeDialog(orderedRowIndices:IOrderedRowIndices, provider: LocalDataProvider): Promise<IExportData> {
-    return import('phovea_ui/dist/components/dialogs').then((dialogs) => {
-      const dialog = new dialogs.FormDialog(`${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.exportData')}`, `<i class="fa fa-download"></i>${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.export')}`);
-      const id = `e${BaseUtils.randomId(3)}`;
-      const ranking = provider.getFirstRanking();
-      dialog.form.classList.add('tdp-ranking-export-form');
-      const flat = ranking.flatColumns;
-      const lookup = new Map(flat.map((d) => <[string, Column]>[d.id, d]));
-      dialog.form.innerHTML = `
-        <div class="form-group">
-          <label>${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.columns')}</label>
-          ${flat.map((col) => `
-            <div class="checkbox tdp-ranking-export-form-handle">
-            <span class="fa fa-sort"></span>
-            <label>
-              <input type="checkbox" name="columns" value="${col.id}" ${!isSupportType(col) ? 'checked' : ''}>
-              ${col.label}
-            </label>
-          </div>
-          `).join('')}
-        </div>
-        <div class="form-group">
-          <label>${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.rows')}</label>
-          <div class="radio"><label><input type="radio" name="rows" value="all" checked>${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.allRows')} (${orderedRowIndices.all.length})</label></div>
-          <div class="radio"><label><input type="radio" name="rows" value="filtered">${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.filteredRows')} (${orderedRowIndices.filtered.length})</label></div>
-          <div class="radio"><label><input type="radio" name="rows" value="selected">${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.selectedRows')} (${orderedRowIndices.selected.length})</label></div>
-        </div>
-        <div class="form-group">
-          <label for="name_${id}">${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.exportName')}</label>
-          <input class="form-control" id="name_${id}" name="name" value="Export" placeholder="${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.nameOfExported')}">
-        </div>
-        <div class="form-group">
-          <label for="type_${id}">${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.exportFormatCapital')}</label>
-          <select class="form-control" id="type_${id}" name="type" required placeholder="${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.exportFormat')}">
-          <option value="csv" selected>${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.csvComma')}</option>
-          <option value="tsv">${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.tsv')}</option>
-          <option value="ssv">${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.csvColon')}</option>
-          <option value="json">${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.json')}</option>
-          <option value="xlsx">${I18nextManager.getInstance().i18n.t('tdp:core.lineup.export.excel')}</option>
-          </select>
-        </div>
-      `;
-
-      ExportUtils.resortAble(<HTMLElement>dialog.form.firstElementChild!, '.checkbox');
-
-      return new Promise<IExportData>((resolve) => {
-        dialog.onSubmit(() => {
-          const data = new FormData(dialog.form);
-
-          dialog.hide();
-
-          const rows = data.get('rows').toString();
-          let order: number[];
-
-          switch(rows) {
-            case 'selected':
-              order = orderedRowIndices.selected;
-              break;
-            case 'filtered':
-              order = orderedRowIndices.filtered;
-              break;
-            default:
-              order = orderedRowIndices.all;
-          }
-
-          const columns: Column[] = data.getAll('columns').map((d) => lookup.get(d.toString()));
-
-          resolve({
-            type: <ExportFormat>data.get('type'),
-            columns,
-            order,
-            name: <string>data.get('name')
-          });
-
-          return false;
-        });
-
-        dialog.show();
-
-        setTimeout(() => {
-          const first = <HTMLElement>dialog.form.querySelector('input, select, textarea');
-          if (first) {
-            first.focus();
-          }
-        }, 250); // till dialog is visible
-      });
     });
   }
